pyproject.toml
[project]
name = "pywebvue-framework"
version = "0.1.0"
description = "Desktop rapid development framework for Python developers (PyWebView + Vue + DaisyUI)"
readme = "README.md"
requires-python = ">=3.10.8"
dependencies = [
    "loguru>=0.7.3",
    "pywebview>=6.1",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pywebvue"]